pyproject.toml
[project]
name = "a2a-datetime-parser-agent"
version = "0.1.0"
description = "An autonomous agent that provides datetime parsing capabilities using the A2A SDK."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk[sql]>=0.3.10",
    "autopep8>=2.3.2",
    "fastmcp>=2.12.5",
    "loguru>=0.7.3",
    "openai>=2.6.1",
    "pydantic>=2.12.3",
    "python-dotenv>=1.1.1",
    "tiktoken>=0.12.0",
    "asyncclick>=8.3.0.7",
    "groq>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "autopep8>=2.0.0",
    "pre-commit>=3.6.0",
]

[project.scripts]

[tool.autopep8]
max_line_length = 120
aggressive = 4

[tool.setuptools.packages.find]
include = ["app*", "cli*"]

[tool.uv]
package = true
